minor tweaks: added class Animal, saving pos.json fixed, remove redundant functions, cleaned logic in move methods of Animal classes
